+++ src/components/Animes/Animes.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-
 import styles from './Anines.module.scss';
 import { AnimeItem } from './AnimeItem/AnimeItem';
-import { TITLE } from '../../const/page-title';
+import { REQUEST_ERROR, TITLE } from '../../const';
 
 interface AnimesProps {
   data: AnimeData[] | undefined;
@@ -14,7 +13,9 @@ export const Animes: React.FC<AnimesProps> = ({ data, open, loading }) => {
   return (
     <div>
       {!loading && (
-        <h2 className={styles.countries_title}>{data?.length === 0 ? 'not found' : TITLE.main}</h2>
+        <h2 className={styles.countries_title}>
+          {data?.length === 0 ? REQUEST_ERROR : TITLE.main}
+        </h2>
       )}
       <div className={styles.countries_container}>
         {data?.map((anime) => (
